[grafos] Arreglada lectura de grafos (los índices son enteros)

+++ src/graph.py
@@ -7,36 +7,36 @@
 
 
 def readGraph ():
-    """ 
+    """
     Formato:
         n
         e
         a b w
         a b w
         [...]
-    
-    Donde "n" es el número de vértices, "e" número de aristas. 
+
+    Donde "n" es el número de vértices, "e" número de aristas.
     Y cada arista es del tipo [a,b] con peso w: (a,b):w
 
     El grafo tendrá nodos [0..n-1], podrá accederse al peso de
     una arista concreta por: graph[(a,b)]
     """
-    
+
     # Define el grafo
     graph = {
         'num_vertices': 0
     }
-    
+
     # Lee dimensión
     graph['num_vertices'] = int(input())
     e = int(input())
-    
+
     # Lee aristas
     for i in range(e):
         a,b,w = raw_input().strip().split()
-        a,b,w = float(a),float(b),float(w)
+        a,b,w = int(a),int(b),float(w)
         graph[(a,b)] = w
-        
+
     return graph
 
 
@@ -45,7 +45,7 @@ def distanceMatrix(graph):
     """ Da la matriz de distancias de un grafo. """
     def dist(a,b):
         # Caso de arista contenida en el grafo
-        if (a,b) in graph: 
+        if (a,b) in graph:
             return graph[(a,b)]
         # Caso simétrico
         elif (b,a) in graph:
@@ -69,4 +69,3 @@ def printMatrix(matrix):
         for j in range(len(matrix[i])):
             print repr(matrix[i][j]).center(3),
         print ')'
-
